pyproject.toml
[project]
name = "traffic-flow-rag"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "llama-index-core>=0.11.8",
    "llama-index-readers-file>=0.2.1",
    "llama-index-llms-ollama>=0.3.3",
    "llama-index-embeddings-huggingface>=0.3.1",
    "llama-stack>=0.0.36",
    "llama-toolchain>=0.0.35",
    "llama-index-llms-huggingface>=0.3.4",
    "llama-index-vector-stores-supabase>=0.2.0",
    "llama-index-llms-lmstudio>=0.2.1",
    "supabase>=2.9.0",
    "python-dotenv>=1.0.1",
    "llama-index-embeddings-openai>=0.2.5",
    "voyageai>=0.2.3",
    "llama-index-embeddings-voyageai>=0.2.2",
    "llama-index-postprocessor-voyageai-rerank>=0.2.0",
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "torchaudio>=2.4.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
rag_llama = 'rag:main'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/traffic_flow_rag"]
